Move Pawn class to separate module
