+++ src/pieces/main.js
@@ -16,117 +16,9 @@ limitations under the License.
 
 
 const { Piece } = require('./base');
+const { Pawn } = require('./pawn');
 const { Relation } = require('../relations');
 const { SquareCoordinates } = require('../square');
-
-
-class Pawn extends Piece {
-    static DIRECTIONS = {[Piece.WHITE]: 1, [Piece.BLACK]: -1};
-    static INITIAL_RANKS = {[Piece.WHITE]: "2", [Piece.BLACK]: "7"};
-
-    constructor(color, square, refresh=true) {
-        /*
-        Params:
-            color {string} (white or black);
-            square {Square} - where piece is placed;
-            refresh {boolean} - whether refresh board after piece placed or not.
-        */
-
-        if (square.onEdge.up || square.onEdge.down) {
-            throw Error(`Pawn couldn't be placed on ${square.name.value} square.`);
-        }
-        super(color, square, Piece.PAWN, refresh);
-    }
-
-    get direction() {
-        return Pawn.DIRECTIONS[this.color];
-    }
-
-    get onInitialRank() {
-        return this.square.onRank(Pawn.INITIAL_RANKS[this.color]);
-    }
-
-    _getMoveCoordinates() {
-        let moveSquaresCoordinates = [];
-        moveSquaresCoordinates.push([
-            this.square.coordinates.x,
-            this.square.coordinates.y + 1 * this.direction
-        ]);
-        if (this.onInitialRank) {
-            moveSquaresCoordinates.push([
-                this.square.coordinates.x,
-                this.square.coordinates.y + 2 * this.direction
-            ]);
-        }
-        return moveSquaresCoordinates;
-    }
-
-    _getMoveSquares() {
-        for (let [x, y] of this._getMoveCoordinates()) {
-            let square = this.board.squares.getFromCoordinates(x, y);
-            if (square.piece) break;
-            this.squares.add(Relation.MOVE, square);
-        }
-    }
-
-    _getAttackCoordinates() {
-        let attackSquaresCoordinates = [];
-        if (!this.square.onEdge.right) {
-            attackSquaresCoordinates.push([
-                this.square.coordinates.x + 1,
-                this.square.coordinates.y + 1 * this.direction
-            ]);
-        }
-        if (!this.square.onEdge.left) {
-            attackSquaresCoordinates.push([
-                this.square.coordinates.x - 1,
-                this.square.coordinates.y + 1 * this.direction
-            ]);
-        }
-        return attackSquaresCoordinates;
-    }
-
-    _checkEnPassantSquare(square) {
-        // Check whether square is en passant square or not
-
-        return (
-            this.board.enPassantSquare
-        &&
-            square.theSame(this.board.enPassantSquare)
-        &&
-            !this.hasColor(this.board.colors.current)
-        );
-    }
-
-    _getAttackSquares() {
-        for (let [x, y] of this._getAttackCoordinates()) {
-            let square = this.board.squares.getFromCoordinates(x, y);
-            this.squares.add(Relation.CONTROL, square);
-            if (square.piece) {
-                if (this.sameColor(square.piece)) {
-                    this.squares.add(Relation.COVER, square);
-                }
-                else {
-                    this.squares.add(Relation.ATTACK, square);
-                    if (square.piece.isKing) {
-                        square.piece.checkers.add(this.square.piece);
-                    }
-                }
-            } else if (this._checkEnPassantSquare(square)) {
-                this.squares.add(Relation.ATTACK, square);
-                this.squares.add(Relation.MOVE, square);
-            }
-        }
-    }
-
-    getSquares() {
-        // Get pawn squares by piece action.
-
-        this._refreshSquares();
-        this._getMoveSquares();
-        this._getAttackSquares();
-    }
-}
 
 
 class StepPiece extends Piece {
